Fix for ODBC-213. Bug inserting SQLINTEGER values in 64 bit Linux

+++ IscDbc/IscHeadSqlVar.h
@@ -64,7 +64,7 @@ public:
 	void		setTypeVarying()	{ MAKEHEAD(SQL_VARYING, 0, 0, 0); }
 	void		setTypeBoolean()	{ MAKEHEAD(SQL_BOOLEAN, 0, 0, sizeof(TYPE_BOOLEAN)); }
 	void		setTypeShort()		{ MAKEHEAD(SQL_SHORT, 0, 0, sizeof (short)); }
-	void		setTypeLong()		{ MAKEHEAD(SQL_LONG, 0, 0, sizeof (long)); }
+	void		setTypeLong()		{ MAKEHEAD(SQL_LONG, 0, 0, sizeof (int)); }
 	void		setTypeFloat()		{ MAKEHEAD(SQL_FLOAT, 0, 0, sizeof (float)); }
 	void		setTypeDouble()		{ MAKEHEAD(SQL_DOUBLE, 0, 0, sizeof (double)); }
 	void		setType_D_Float()	{ MAKEHEAD(SQL_D_FLOAT, 0, 0, sizeof (double)); }
